pyproject.toml
[tool.poetry]
name = "securitization-comparison"
version = "0.1.0"
description = "Compare traditional securitization with tokenized/blockchain-based securitization on real-world mortgage data"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "securitization_comparison"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
plotly = "^5.15.0"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
pyyaml = "^6.0.1"
streamlit = "^1.28.0"
scipy = "^1.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
ruff = "^0.0.280"
black = "^23.7.0"
mypy = "^1.5.0"
types-pyyaml = "^6.0.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "D", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "RUF"]
ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true 